requirements/dev.txt
Django==1.6.1
Django-tinymce-filebrowser==0.2.1
Pillow==2.3.0
South==0.8.4
Unipath==1.0
coverage==3.7.1
django-admin-tools==0.5.1
django-mockups==0.4.8
django-taggit==0.11.2
django-tinymce==1.5.2
httplib2==0.8
pip-tools==0.3.4
psycopg2==2.5.2
python-openid==2.2.5
six==1.5.2
sorl-thumbnail==11.12
wsgiref==0.1.2
